little changed logic in admin and blog controllers
