+++ src/main/java/com/example/servingwebcontent/model/entity/Role.java
@@ -20,8 +20,9 @@ public enum Role {
     }
 
     public Set<SimpleGrantedAuthority> getAuthorities(){
-        return getPermissions().stream()
+        Set<SimpleGrantedAuthority> authorities = getPermissions().stream()
                 .map(permission -> new SimpleGrantedAuthority(permission.getPermission()))
                 .collect(Collectors.toSet());
+        return authorities;
     }
 }
