+++ src/main/java/com/example/servingwebcontent/security/SecurityUser.java
@@ -7,6 +7,7 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 
@@ -61,12 +62,15 @@ public class SecurityUser implements UserDetails {
     }
 
     public static UserDetails fromUser(User user){
+        var authorities = new ArrayList<GrantedAuthority>();
+        user.getRole().forEach(role -> authorities.add(new SimpleGrantedAuthority(role.toString())));
+
         return new org.springframework.security.core.userdetails.User(
                 user.getUsername(), user.getPassword(),
                 user.getStatus().equals(Status.ACTIVE),
                 user.getStatus().equals(Status.ACTIVE),
                 user.getStatus().equals(Status.ACTIVE),
                 user.getStatus().equals(Status.ACTIVE),
-                user.getRole().stream().iterator().next().getAuthorities());
+                authorities);
     }
 }
