+++ src/main/resources/templates/blocks/header.html
@@ -3,7 +3,7 @@
     <!-- As a heading -->
     <!DOCTYPE HTML>
     <html xmlns:th="http://www.thymeleaf.org"
-          xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity4">
+          xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity5">
     <head>
         <title th:text="${title}"/>
         <meta charset="utf-8">
@@ -27,7 +27,7 @@
                     <li class="nav-item active">
                         <a class="nav-link" href="/">Main page<span class="sr-only"></span></a>
                     </li>
-                    <li class="nav-item">
+                    <li class="nav-item" sec:authorize="!isAuthenticated()">
                         <a class="nav-link" href="/login">Personal account</a>
                     </li>
 
@@ -36,11 +36,11 @@
                     </li>
 
                     <li class="nav-item" sec:authorize="isAuthenticated()">
-                        <a class="nav-link" href="/user/profile">Profile</a>
+                        <a class="nav-link" href="/blog/profile">Profile</a>
                     </li>
 
-                    <li class="nav-item" >
-                        <a class="nav-link" th:href="@{/user}">User list</a>
+                    <li class="nav-item" sec:authorize="hasRole('ADMIN')">
+                        <a class="nav-link" href="/admin">User list</a>
                     </li>
 
 
@@ -48,9 +48,9 @@
                         <a class="nav-link" href="/about">About us</a>
                     </li>
                 </ul>
-                <form class="form-inline my-2 my-lg-0">
-                    <input class="form-control mr-sm-2" type="search" placeholder="Поиск" aria-label="Поиск">
-                    <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
+
+                <form class="form-inline my-2 my-lg-0" th:action="@{/logout}" method="post" sec:authorize="isAuthenticated()">
+                        <input type="submit" value="Logout"/>
                 </form>
             </div>
             </nav>
