+++ src/main/resources/templates/blog-main.html
@@ -40,10 +40,6 @@
 </div>
 
 
-<div>
-    <form action="/logout" method="post">
-        <input type="submit" value="Выйти"/>
-    </form>
-</div>
+
 
 <div th:insert="blocks/footer :: footer"></div>
